+++ src/helpers.ts
@@ -1,3 +1,5 @@
+import { readFileSync } from "node:fs";
+import { join as joinPath } from "node:path";
 import { flow, join, split, startsWith, take } from "ramda";
 
 /** is scoped import: starts with "at" */
@@ -6,3 +8,6 @@ const hasScope = startsWith("@");
 /** gets the dependency name even when importing its internal path */
 export const getName = (imp: string) =>
   flow(imp, [split("/"), take(hasScope(imp) ? 2 : 1), join("/")]);
+
+export const getManifest = (path: string) =>
+  JSON.parse(readFileSync(joinPath(path, "package.json"), "utf8"));